Move Link to boilerplate as LinkListEntry.svelte
