+++ src/lib/boilerplate.ts
@@ -1,6 +1,7 @@
-import Who from './boilerplate/Who.svelte.md';
-import Socials from './boilerplate/Socials.svelte';
 import Donate from './boilerplate/Donate.svelte.md';
+import LinkListEntry from './boilerplate/LinkListEntry.svelte';
+import Socials from './boilerplate/Socials.svelte';
+import Who from './boilerplate/Who.svelte.md';
 import WwpMusic from './boilerplate/WwpMusic.md';
 
-export { Who, Socials, Donate, WwpMusic };
+export { Donate, LinkListEntry, Socials, Who, WwpMusic };
